+++ test/support/unit.ex
@@ -2,8 +2,8 @@ defmodule Test.Support.Unit do
   import ExUnit.Assertions
 
   def assert_format_string(bad, good) do
-    assert ExFormat.format_string(bad) == good
-    assert ExFormat.format_string(good) == good
+    assert ExFormat.format(bad) == good
+    assert ExFormat.format(good) == good
   end
 
   def assert_format_string(string) do
